+++ guidata/widgets/console/__init__.py
@@ -18,7 +18,7 @@ This package provides a Python console widget.
 
 from qtpy.QtCore import Qt
 
-from guidata.config import CONF, _
+from guidata.config import CONF
 from guidata.configtools import get_font
 from guidata.qthelpers import win32_fix_title_bar_background
 from guidata.widgets.console.internalshell import InternalShell